Create db in SUSPECT state
With this article you will be able to investigate d banks ed corrupted as well as Corrig i them using the tools that SQL Server and Windows offer. In it, we will define some common mistakes and also assist in decision making to reduce downtime in case the database goes into “ SUSPECT ” state . For this, we will work on two practical examples : one with the database with the corrupted log and the other without the data log (typical when the file has been completely lost due to hardware or software failure). With this practice, we will know how to react when the bank is in the “SUSPECT” state .