Why this post is useful
With this article we will analyze how the fragmentation of windows can impact the physical files of the databases in SQL Server. We will analyze what procedures we can take so that windows defragmentation has an effect on these files in addition to observing their performance within the database. We will study the structure of a file inside windows and how SQL Server behaves to record information within the datafile and how it organizes this information internally.